backend/requirements/base.txt
Django==5.1.6
daphne==4.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
psycopg[binary]==3.2.3
redis==5.2.1
celery==5.4.0
django-cors-headers==4.6.0
channels==4.2.0
channels-redis==4.2.1
python-dotenv==1.0.1
structlog==24.4.0
orjson==3.10.12
pydantic==2.10.4
pydantic-settings==2.7.0
networkx==3.4.2
pgvector==0.3.6
